correcção de pequenos erros

+++ test/Dominio/PortaTest.java
@@ -39,9 +39,9 @@ public class PortaTest {
      */
     @Test
     public void testCalculaArea() {
-        double altura = 2.5;
-        double largura = 4.2;
-        Porta porta = new Porta(altura, largura, 1.5);
+        float altura = (float) 2.5;
+        float largura = (float) 4.2;
+        Porta porta = new Porta(altura, largura, (float) 1.5);
         double resultado = 10.5;
         assertEquals("CalculaAreaTest", resultado, porta.calculaArea(), 0.0);
     }
